[ITQT1-26] Added api to get report by id
