+++ apps/api/src/routes/quiz.route.ts
@@ -3,8 +3,8 @@ import {
     createQuiz,
     getAllQuiz,
     getQUiz,
-    getReports,
 } from '../controllers/quiz.controller';
+import { getReports } from '../controllers/report.controller';
 
 const router = express.Router();
 
@@ -12,6 +12,7 @@ router.route('/').post(createQuiz).get(getAllQuiz);
 
 router.route('/report').get(getReports);
 
+// * NOTE: This should be at last as it catches other routes also.
 router.route('/:id').get(getQUiz);
 
 export default router;
